More slider fixes 2 (#1037)
* more fixes

* more fixes
