+++ packages/webawesome/docs/docs/components/slider.md
